requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
alsa-lib=1.2.3=h516909a_0
argon2-cffi=21.1.0=py39h3811e60_2
asgiref=3.4.1=pyhd8ed1ab_0
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bleach=4.1.0=pyhd8ed1ab_0
boost-cpp=1.74.0=h312852a_4
branca=0.4.2=pyhd8ed1ab_0
brotli=1.0.9=h7f98852_6
brotli-bin=1.0.9=h7f98852_6
brotlipy=0.7.0=py39h3811e60_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2021.10.8=py39hf3d152e_1
cffi=1.15.0=py39h4bc2ebd_0
cfitsio=4.0.0=h9a35b8e_0
chardet=4.0.0=py39hf3d152e_2
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.3=py39hf3d152e_1
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
cryptography=36.0.0=py39h95dcef6_0
curl=7.80.0=h2574ce0_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h48d8840_2
debugpy=1.5.1=py39he80948d_0
decorator=5.1.0=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
django=3.2.9=pyhd8ed1ab_0
django-extensions=3.1.5=pyhd8ed1ab_0
django-model-utils=4.0.0=py_0
django-rest-swagger=2.2.0=py_0
djangorestframework=3.12.4=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
et_xmlfile=1.0.1=py_1001
expat=2.4.1=h9c3ff4c_0
fiona=1.8.20=py39h427c1bf_2
folium=0.12.1.post1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.1=hba837de_1005
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.28.1=py39h3811e60_0
freetype=2.10.4=h0708190_1
freexl=1.0.6=h7f98852_0
gdal=3.3.3=py39h0494519_2
geopandas=0.10.2=pyhd8ed1ab_0
geopandas-base=0.10.2=pyha770c72_0
geos=3.10.0=h9c3ff4c_0
geotiff=1.7.0=hcfb7246_3
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
glib=2.70.1=h780b84a_0
glib-tools=2.70.1=h780b84a_0
gst-plugins-base=1.18.5=hf529b03_2
gstreamer=1.18.5=h9f60fe5_2
hdf4=4.2.15=h10796ff_3
hdf5=1.12.1=nompi_h2750804_102
httpie=2.1.0=py39hf3d152e_2
icu=68.2=h9c3ff4c_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.8.2=py39hf3d152e_0
importlib_resources=5.4.0=pyhd8ed1ab_0
ipykernel=6.5.0=py39hef51801_1
ipython=7.29.0=py39hef51801_2
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.5=pyhd8ed1ab_0
itypes=1.2.0=pyh9f0ad1d_0
jbig=2.1=h7f98852_2003
jedi=0.18.1=py39hf3d152e_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
json-c=0.15=h98cffda_0
jsonfield=3.1.0=py_0
jsonschema=4.2.1=pyhd8ed1ab_0
jupyter_client=7.1.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py39hf3d152e_1
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.2=pyhd8ed1ab_0
kealib=1.4.14=h87e4c3c_3
kiwisolver=1.3.2=py39h1a9c180_1
krb5=1.19.2=hcc1bbae_3
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_openblas
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libcblas=3.9.0=12_linux64_openblas
libclang=11.1.0=default_ha53f305_1
libcurl=7.80.0=h2574ce0_0
libdap4=3.20.6=hd7c4107_2
libdeflate=1.8=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_11
libgdal=3.3.3=h18e3bf0_2
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libglib=2.70.1=h174f98d_0
libgomp=11.2.0=h1d223b6_11
libiconv=1.16=h516909a_0
libkml=1.3.0=h238a007_1014
liblapack=3.9.0=12_linux64_openblas
libllvm11=11.1.0=hf817b99_2
libnetcdf=4.8.1=nompi_hb3fd0d9_101
libnghttp2=1.43.0=h812cca2_1
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.18=pthreads_h8fe5266_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.5=hd57d9b9_0
librttopo=1.1.0=h0ad649c_7
libsodium=1.0.18=h36c2ea0_1
libspatialindex=1.9.3=h9c3ff4c_4
libspatialite=5.0.1=h1d9e4f1_10
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.3.0=h6f004c6_2
libuuid=2.32.1=h7f98852_1000
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzip=1.8.0=h4de3113_1
libzlib=1.2.11=h36c2ea0_1013
lz4-c=1.9.3=h9c3ff4c_1
mapclassify=2.4.3=pyhd8ed1ab_0
markupsafe=2.0.1=py39h3811e60_1
matplotlib=3.5.0=py39hf3d152e_0
matplotlib-base=3.5.0=py39h2fa2bec_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py39h3811e60_1005
munch=2.5.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.27=ha770c72_1
mysql-libs=8.0.27=hfa10184_1
nbclient=0.5.9=pyhd8ed1ab_0
nbconvert=6.3.0=py39hf3d152e_1
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
networkx=2.6.3=pyhd8ed1ab_1
notebook=6.4.6=pyha770c72_0
nspr=4.32=h9c3ff4c_1
nss=3.72=hb5efdd6_0
numpy=1.21.4=py39hdbf815f_0
olefile=0.46=pyh9f0ad1d_1
openapi-codec=1.3.2=py_0
opencp=0.2.0=pypi_0
openjpeg=2.4.0=hb52868f_1
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1l=h7f98852_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.4=py39hde0f152_1
pandoc=2.16.2=h7f98852_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.2=pyhd8ed1ab_0
patsy=0.5.2=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py39ha612740_0
pip=21.3.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
plotly=5.4.0=pyhd8ed1ab_0
poppler=21.09.0=ha39eefc_3
poppler-data=0.4.11=hd8ed1ab_0
postgresql=13.5=h2510834_0
proj=8.1.1=h277dcde_2
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.22=pyha770c72_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.10.0=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pyproj=3.2.1=py39ha81a305_2
pyqt=5.12.3=py39hf3d152e_8
pyqt-impl=5.12.3=py39hde8b62d_8
pyqt5-sip=4.19.18=py39he80948d_8
pyqtchart=5.12=py39h0fcd23e_8
pyqtwebengine=5.12.1=py39h0fcd23e_8
pyrsistent=0.18.0=py39h3811e60_0
pysocks=1.7.1=py39hf3d152e_4
python=3.9.7=hb7a2778_3_cpython
python-coreapi=2.3.3=py_0
python-coreschema=0.0.4=py_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd8ed1ab_0
pyzmq=22.3.0=py39h37b5a0c_1
qt=5.12.9=hda022c4_4
readline=8.1=h46c0cb4_0
requests=2.26.0=pyhd8ed1ab_0
rtree=0.9.7=py39hb102c33_3
scikit-learn=1.0.1=py39h4dfa638_2
scipy=1.7.2=py39hee8e79c_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=59.2.0=py39hf3d152e_0
shapely=1.8.0=py39hc7dd4e9_2
simplejson=3.17.6=py39h3811e60_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h9cd32fc_2
sqlparse=0.4.2=pyhd8ed1ab_0
statsmodels=0.13.1=py39hce5d2b2_0
tenacity=8.0.1=pyhd8ed1ab_0
terminado=0.12.1=py39hf3d152e_1
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=3.0.0=pyh8a188c0_0
tiledb=2.3.4=he87e0bf_0
tk=8.6.11=h27826a3_1
tornado=6.1=py39h3811e60_2
traitlets=5.1.1=pyhd8ed1ab_0
typing_extensions=4.0.0=pyha770c72_0
tzcode=2021e=h7f98852_0
tzdata=2021e=he74cb21_0
uritemplate=4.1.1=pyhd8ed1ab_0
urllib3=1.26.7=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.0=pyhd8ed1ab_1
widgetsnbextension=3.5.2=py39hf3d152e_1
xerces-c=3.2.3=h9d8b166_3
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xyzservices=2021.11.0=pyhd8ed1ab_0
xz=5.2.5=h516909a_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.0=ha95c52a_0
